requirements.txt
django==1.6.3
south==0.8.4
celery-with-redis==3.0
django-celery==3.1.10
psycopg2==2.5.2
redis==2.9.1
python-memcached==1.53